packages/app/pyproject.toml
[project]
name = "spend-app"
dynamic = ["version"]
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.119.0",
    "pydantic-settings>=2.11.0",
    "uvicorn>=0.37.0",
    "databricks-sdk>=0.74.0",
    "sqlmodel>=0.0.27",
    "psycopg[binary,pool]>=3.2.11",
    "spend-categorization",
    "pandas>=2.0.0",
    "pyyaml>=6.0.0",
    "psycopg2-binary>=2.9.0",
]

[tool.uv.sources]
spend-categorization = { workspace = true }

[dependency-groups]
dev = [
    "ty>=0.0.12",
]

[tool.apx.metadata]
app-name = "spend-app"
app-slug = "spend_app"
app-entrypoint = "spend_app.backend.app:app"
api-prefix = "/api"
metadata-path = "src/spend_app/_metadata.py"

[tool.apx.ui]
root = "src/spend_app/ui"

[tool.apx.ui.registries]
"@animate-ui" = "https://animate-ui.com/r/{name}.json"
"@ai-elements" = "https://registry.ai-sdk.dev/{name}.json"
"@svgl" = "https://svgl.app/r/{name}.json"

[build-system]
requires = ["hatchling", "uv-dynamic-versioning>=0.7.0"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.uv-dynamic-versioning]
vcs = "git"
fallback-version = "0.0.0"

[tool.hatch.build.hooks.version]
path = "src/spend_app/_version.py"
template = '''
version = "{version}"
'''

[tool.hatch.build]
artifacts = ["src/spend_app/__dist__", "src/spend_app/_metadata.py"]
exclude = ["src/spend_app/ui"]
